Add admonition to telemetry code
This code is subject to change in the near future.  The alarm
class is probably moving out of here and many of the resource
property have aliases.  We have agreed not to support the old
aliases going forward, because it doesn't work well.  The
admonition can be pulled after these issues are fixed.

Closes-Bug: #1468086
Change-Id: Ica464a6c1e2281f4acbd25a121a399b6b0c81376