Get the status of the ip with ip.get('status')
Ensure that there is no error if the key doesn't exist in the ip.

The API spec indicates this field should be a string and should always
exist. However, it has been reported in the wild that sometimes it's
missing. Yay for specs. If there is no status, set it to 'UNKNOWN' so
that there is at least some indication in logging for a user as to what's
going on.

Change-Id: I27a8bd35e6614ac24b8cbb1ada8c101329fd2f7f