Support block_store types where IDs are taken
In an effort to widen support for resource properties that are IDs but
can take the appropriate resource type for that ID, the block_store
service throws a kink into it. The Volume type can contain a Snapshot,
and a Snapshot can contain a Volume. Because of that loop, we can't do
the usual `foo = resource.prop(foo, FooType)`. We can sort of emulate
the behavior users will get out of it in the proxy by doing a type
conversion in the create_volume and create_snapshot methods to support
resource instances being passed in there.

Additionally, Volume can be created from another Volume, but we can't
have a prop refer to its own class, so that problem is solved the same
as above.

Lastly, the Volume resource has two props that can take real types:
image and type.

Change-Id: If8a5b24b05a341ae6bfe9dca6b93429dcfc8f408