Skip orchestration functional tests
There no longer seems to be enough resources on the gate to
consistently create a heat stack with one instance *and* a
run compute.server_create. It is more important to test compute
than orchestration, so disable those tests for now and fix
orchestration or delete the test in the future.

Bug to track:
https://bugs.launchpad.net/python-openstacksdk/+bug/1525005

Change-Id: Ie86329155fffeef2a2f9f6c9927295380dc72671