Fix "report a bug" launchpad project
For docs set the launchpad project users report bugs
against to "python-openstacksdk". Users can report bugs
using the "bug icon" that will
directly link to the launchpad project, it currently goes to
"openstack-manuals" which is wrong for this content.

This variable is used by openstackdocstheme 1.2.6.

Also, update comments for the variables passed to openstackdocstheme.

Change-Id: Id036c6373092673ce0c54be1b64ea87d1de89b88
Related-Bug: #1524476