Remove requests from requirements
With the move to keystoneauth, we no longert depend on the requests
library for anything at runtime. We do still have one test, which is
valid, that depends on requests to ensure that
https://bugs.launchpad.net/python-openstacksdk/+bug/1428342 stays fixed.
This change moves requests to test-requirements and additionally removes
the list of dependencies in the README file, as it duplicates the only
place requirements are truly needed.

Change-Id: I6aa8e71573e685b2748150ea5cec946b65d525ae