Add support for querying role assignments.
When you are granting or revoking roles, you need to know what the role
assignments are. Also, the return dict for role_assignments is
exceptionally hard to query, so turn it in to a structure that makes
some sanity.

Note that role assignments don't lend themselves to use our list/search/get
standard very well. Assignments are not unique (an assignment has neither a
name nor an ID). Searching is supported via the filter push-down to the
keystone client itself.

This replaces: I963b7a07cb711d3f790715a1fe55b16d0200073d

Change-Id: I5fb96f32a2f1a5f0408938105750afcd7c5c93af
Co-Authored-By: Monty Taylor <mordred@inaugust.com>