Return None when getting an attr which is None when using resource.prop()
In resource.prop, __get__() will do self.type() on the value it got to
construct an object of the type. But the type could be overwritten to any
type. And the type may not be able to construct an object from None.

For example, if self.type=dict, then self.type(None) will become dict(None),
which will give "'NoneType' object is not iterable" error.

This patch fix it by returning None if the attr's value is None.

Change-Id: Ied255c52bc3f0d024ca70d1239956ba17430b13e
Closed-bug: #1527967