Fix glance endpoints with endpoint_override
Now that we properly pass endpoint_override all the time, we broke
glance. The reason for this is that we calculate the glance url via
glance url stripping in all cases, so the case where we did not have
a configured endpoint override was passing the wrong information to the
constructor, causing double version addition.

Change-Id: I5699b0581d0cb68fed68800c29c8a847e2606ec9