If cloud doesn't list regions expand passed name
Don't fail on a cloud not having regions when a region name is passed.
Instead just use the name that is given and expand it properly.

This adds test coverage for the paths through the
OpenStackConfig._get_region() method to avoid problems like this in the
future.

In order for this work to be done cleanly a small refactor of
get_regions() is done to split it into two methods, one that gets all
regions with a sane fallback default (for backward compat) and another
that returns only regions that are known in the config and None
otherwise. This allows us to switch on whether or not there are known
regions.

Change-Id: I62736ea82f365badaea5016a23d37a9f1c760927