Munge region_name to '' if set to None
The openstack ansible module defaults to setting region_name to
None[1]. With region_name explicitly set, _get_region won't use '' as a
default and therefor has unexpected behavior if the user does not set
the region explicitly. This is apparent in bifrost[2] which does not
use any cloud config file and does not set the region explicitly. This
patch checks whether None was passed in as the region name and sets it
to '' so that it can continue processing it as though it was not set.

[1] https://github.com/ansible/ansible/blob/devel/lib/ansible/module_utils/openstack.py#L41
[2] http://paste.openstack.org/show/482831/

Change-Id: I22cce104930f74dd479e704cc1a941dc945b75de