Normalize server objects
There are a set of things we always want to do to server dicts because
they are no-cost. This is in line with how we manage other resource
dicts, so make a normalize function. However, because we need to add
cloud and region info from the cloud object, make it a member method
rather than an friend function. Then, ensure it's always called on
server related calls.

Change-Id: Ied869efcb7d83c4f1a753962ca3127ce52172a24