Add range search functionality
It can be useful to have more advanced search capability across a
data set than just exact value matching. For instance, when searching
for flavors, one may want to just say "give me the flavor(s) with the least
amount of ram" or "give me the flavor(s) with 2 or more virtual CPUs, but less
than 8GB of ram". We can accomplish this with range searching. Example:

    flavors = self.range_search(self.list_flavors(),
                                {"vcpus": ">=2", "ram": "<8192"})

Besides the normal range operators (<, >, <=, >=), one can also use
the values "MIN" or "MAX" for minimum and maximum values:

    flavors = self.range_search(self.list_flavors(), {"ram": "MIN"})

Change-Id: I706e4eee62a969888db3ea70f7052d3fb00c544e