Send keystoneauth a better user-agent string
When we moved to keystoneauth sessions from our own handwritten session
class, we lost the nice user-agent string we built up. That was then
added to keystoneauth and has now been released in version 2.2.0, which
we picked up. That implementation provides a good base for us to build
on, and we should now send our own version string as well as make way
for callers to submit their application's name/version string as well.

Change-Id: I439b0912b27a37c4ddfc91447a5f3394df2d0385