Add UNIXEpoch formatter as a type for properties
Several properties in the object_store service use timestamp values that
are a count of seconds since the UNIX Epoch, or January 1, 1970. This
change adds a formatter that exposes this value to the user as a
datetime object, similar to the ISO8601 formatter. It is currently in
use by the transaction timestamp X-Timestamp header on all object_store
resources as well as the X-Delete-At header value on objects.

Change-Id: I3a14e3fae0e22d1e1ae411a5b119a08c6602c2b9