Remove 'date' from Object resource
While creating the various type formatters for properties, I created one
for the timestamp format used on the Date prop that is included in the
headers of account, container, and object resources. However, that
property just duplicates that of the X-Timestamp prop. Since the format
of X-Timestamp is also used elsewhere, it doesn't make much sense to
both include the duplicated info in Date along with having a custom
formatter for it. This change removes Date entirely (it never existed on
account or container, but was about to be added in another change which
spurred its removal here).

Change-Id: Ia458f0d6cb4ba4f4500df97cd58c62afd6d5b05f