Modify test workaround for extra_dhcp_opts
The neutron create API now returns extra_dhcp_opts. In older versions
it did not so it had to be removed when comparing the port definitions
between the "create" and the "get". Now we only remove it if it is not
present in the "create". This keeps it working for our keystonev2 test
job which runs against stable/liberty now.

Change-Id: I4d6a57b0f15f1868c5ba4e62a2bf248662bd8896