Use network in neutron_available_floating_ips
The old behavior takes a list of network names or id, ignores it and
picks the first external network it can find

This fix implements the actual logic of using the parameter:
 * Uses the given external network, if supplied. This changes to
   using only a single network instead of a list because no where
   is a list supplied to the call.
 * If no network is given, it takes the first external network.

(Shrews) The unit tests are also changed because they were suck.
The test for available_floating_ip() is changed to test only it's
calls and not the lower level functionality. New tests are added for
the _neutron_available_floating_ips() call, which implements the lower
level functionality and which is where this change is actually tested.

Co-Authored-By: David Shrewsbury <shrewsbury.dave@gmail.com>

Change-Id: I0b27b50b2a6a8e5199bbeed5786b71851bb3ad3e