Refactor Resource to better serve Proxy
Through a series of attempts we haven't been able to sufficiently
solve the issue of what were called "path_args" in the past,
somewhat of a relic of the time before we had the Proxy layer.
"path_args" made a lot of sense back when there was only the
Resource layer, when the SDK's usage was almost entirely through
the classmethods on Resource, back when command-line example
scripts were all we had to go on.

As things have grown in the year-plus since that time, we have
bolted a great higher-level interface on a lower-level interface
that was stuck in an older use-case. It was most recently a problem
when we had resource.prop types that accepeted Resource types,
which both there and elsewhere meant it became significantly harder -
if not impossible - to have nested Resources. An example of this is
how ServerInterface is a Resource that depends on Server.

This new approach strips Resource down to a public API of only create,
get, update, delete, head, find, and list methods. Resource subclasses
are still constructed in the same manner, though the naming of what
were "props" is now slightly different, but more clear. Additionally,
through that change comes a more clear user experience for URI
parameters, formerly known as, "path_args". The changes here pave the
way for more clarity on query parameters, although they haven't been
fully explored.

As the changing of this class has pretty far reaching effects, it will
exist temporarily in the "resource2" module. The first change will be
to introduce the module itself as a standalone, and then we will change
each service independent of the others. Subsequent changes should not
be too disruptive, mostly in the form of renaming things like
resource.prop to resource.Body, and the like.

There is a fairly major side effect of the implementation of this class
that is behaviorally different from the initial Resource
implementation. Previously, any-and-all data sent into a Resource
initializer would be tracked internally on the Resource instance and
then sent to the server *even if it did not correspond to a
resource.prop on the class*. Ultimately this would have become a
problem as servers are becoming more strict about what they receive,
but this implementation makes no allowance for so-called loose data
to be entered. If there is no component attribute for an argument
to be sent to the server, or one received from the server, it will
not exist as an attribute of the Resource. Currently, said data is
just dropped entirely, but it could be logged or perhaps a
strict/loose mode could be engaged to warn or otherwise fail when
unexpected data is entered. Ultimately, this behavior had to happen.
In the near term it may mean that the SDK doesn't work exactly as it
had in previous versions, but that will just mean updating Resource
definitions to match what is truly expected or given by the server.

A relatively major change that comes from this refactoring is a
deconstruction of the unit tests that previously existed.
While resource.prop now lives on as resource2._BaseComponent,
the unit tests were started over from scratch as they were too
tightly coupled to the Resource class rather than being tested as
an independent unit. All of the code that is being refactored
as a part of this change is now being tested as independent of
any other code as is possible.

Additional changes:

1. Removed support for aliases in _BaseComponent.

Aliased property names were added for samples coming out of the
telemetry service as they used to be supported and documented having
multiple names. However, they are no longer documented this way.

Additionally, aliases are abused in object_store due to a
misunderstanding which causes actual breakage during use. One
resource.prop was in some cases serving to retreieve both header
and body values with similar names. Ultimately, aliases for
property names are not needed and their support is removed
as we have separated out the storage of body and header components.

2. Query parameters are now more officially and easily supported through
the _query_mapping attribute on Resource2. The resource2.QueryParameters
class creates the appropriate mapping of client-side names to
server-side names, which is then used when calling Resource2.list in
order to transpose the names and make a proper request.

3. As all tests were rewritten from scratch, tests for the wait_for*
functions end up being 8-10 times faster as they mock out time.sleep
rather than actually waiting, thus making test runs a lot quicker.

Change-Id: I634ac93b913e4fe9a3a812b384c8539d9266f2c0