Add ability to pass just filename to create_image
There is a common use case of wanting to create a disk image and then
upload it to the cloud where the name of the diskimage minus the file
extension is the desired name of the image in the cloud. Support that
easily by checking to see if the user passed in a filename as the first
paramater and, if so, use it as the filename and strip the path and
extension to use as the name.

Also, document the method while we're in there.

Change-Id: Iade5f071544775864534660b544a6a01b15a2adb