Fixes for latest cinder and neutron clients
In obj_to_dict, if we get a dict-like-object, we should convert that
with Munch. Otherwise, we just convert its properties and not values.

When comparing cinder things, do not compare the entire structure
since what we get from the create call can be different than what
we get from the list call. Just compare IDs.

Change-Id: Iada5ff76cf9192a2c10e117536546989633c290e
Co-Authored-By: Monty Taylor <mordred@inaugust.com>
Co-Authored-By: Sam Yaple <sam@yaple.net>