Use warlock in the glance v2 tests
We were mocking what should have been a warlock object with a dict like
object. Instead of doing that, actually pull the model from glanceclient
and construct a legit warlock object in the mock so that we can make
sure our warlock morphing does the right thing. Also, warlock triggers
'smarts' about which parameters to update, so update the test to mock
out the right things.

Sadly we have to copy the task schema in, because the only place it
exists in API form is in the glance source tree.

Change-Id: I9a63bfb7a85e69e66d32cf28d0e7fe207996e1b4