Bug fix: Do not fail on routers with no ext gw
Although I've not been able to reproduce it, some user have reported
an exception from shade in the list_router_interfaces() call when
trying to access the external_gateway_info of a router that does not
have this key set. Let's just be safe and and a check to make sure
that the key exists.

Change-Id: I949b76b2b306e5161e7ee77d6c588a77ac4c7d87