Pass specific cloud to openstack_clouds function
The openstack_clouds function gets all the clouds defined in
the clouds.yaml.
For clouds with multiple regions, it will return an OpenStackCloud
object for each one of the regions tied to the cloud.
This change allows to pass an specific cloud in case the user
wants to pull a single clouds with multiple regions from clouds.yaml.

Change-Id: I41512825cac99fae80c61b3a61248442dcd637c7