Deal with is_public and ephemeral in normalize_flavors
There are two attributes that are provided by extensions that we should
deal with in our normalization function. novaclient provides helper
methods to make the names less suck. We should also include the suck
versions of the names though, as they are currently public parts of the
API and people might be confused if they're missing.

Change-Id: Ia2f1baf1307ec97fb0166cbb4a633852b461770b