BaseProxy refactoring for new Resource
To work with the changes in https://review.openstack.org/#/c/289998/
there are some follow-on changes necessary in the BaseProxy. This is
primarily the removal of `path_args` arguments from BaseProxy
signatures in favor of using keyword arguments, and passing those
keyword arguments down into Resource methods as well.

As with the Resource refactoring, as this is a major undertaking when it
comes to the entirety of the services affected, this is going to
temporarily live in a secondary "proxy2" module so that we can apply the
changes in waves. There is one temporary change in
openstack/connection.py that allows us to load proxy classes that are
subclassed from openstack.proxy2. The rest of it will eventually live on
as the only proxy class once we have completed all services being
transitioned onto the new one.

Change-Id: I918a31078f157354b5ccad313c51ace50241ede4