Fix test_get_server_ip unit test
It turns out this unit test wasn't actually testing get_server_ip(),
and the things it wanted to test were already being tested by other
unit tests. And because it didn't do mocking properly, unlike the other
tests, it was taking a long time to run. Rework the test to be more
betterer and do correct things.

Change-Id: Ife0fa7673081c72ca766e8f78ad2d44b1d0035f8