Workaround multiple private network ports
In the case where there are multiple private networks, don't actively
throw an error. Instead, just grab the first port and attach NAT to it.
That's not completely correct, so a follow up patch is forthcoming that
will add support to os-client-config for such a configuration value.
However, in the mean time, just grab one, because why not.

Change-Id: Ib3a49c0efc72a6de0bdd84c3fe1f34cc5f553a8c