Cache ports like servers
This implements nearly the same caching logic for ports as servers
so that when nodepool has a lot of servers that need floating ips,
we can potentially list all of the ports and get the ports for many
servers with one API call.

If a filter is being pushed down to neutron, we will just perform
the API call rather than caching (but in the common case of searching
for a port for a server, we will no longer push down the filter
in order to allow the caching to happen).

Since we could have a caching delay, we also wait a little bit
in case we don't get a port when we expect one.

Change-Id: I66566f7a84d50b562f1303ed0e645cb41b11367e