Search subnets for gateway_ip to discover NAT dest
The network that should be the target of the NAT of a floatingip should
have a subnet that has a gateway_ip. Ignore for a moment the
monumental insanity which is that data model ... that a network can have
one or more subnets, and a network_id is what a port has and is how you
describe what you want to attach to, but even with all of that the
gateway_ip is the cogent property and it's needs to be on all of the
subnets in the network. Just ignore that ... if you can.

Since that is the world we live in, we can discover which network is the
netowrk that has one or more subnets that has the gateway_ip set.

We'll still fail if there are more than one network that has a subnet
with gateway_ip set. But in that case, the config will take care of it.

Change-Id: Ib947081a6b7839e1a6f00dcbfd924cc99f19d0fd