Properly handle overridden Body properties
The Resource._get_mapping method was previously overwriting mapping
entries as it traversed the MRO, so while it would properly map
name=nombre on the first pass through the class it's trying to find, it
would then look at base classes, including the base Resource, and then
overwrite that mapping to go back to name=name. This was just seen as an
issue while applying the refactoring to AvailabilityZone in compute,
where the name=zoneName mapping wasn't properly working. This change
fixes that.

Change-Id: I21f343fdbf3c6d5696e3f074d56d96b4fe4faaba