Fix Resource.list usage of limit and marker params
The Great Refactoring brought about a change to how query parameters
were passed in, and while changing Resource.list, support for said query
paramters was added but wasn't accurately propogated throughout the
method. The session.get "params" parameter was being sent the contents
of a transposed mapping of query params the resource takes, but runs
through the "while more_data" loop were setting limit and marker values
in a different dictionary. This removed our ability to properly
paginate real response data because we were just making the same request
over and over. The change was simply to set the limit and marker values
on this new `query_params` dictionary, which is what is sent to
session.get.

The bigger change here is more accurate testing of what is actually
happening in these list method calls.

Change-Id: I0084424f9a8cd97d133f8779e7ace8b8f64a0c24