Add public helper method for cleaning floating ips
Some things, like nodepool or utility scripts, need to be able to safely
clean floating ips, but doing so requires knowledge of whether the cloud
in question is using nova or neutron for floating ips, which shade
otherwise hides from the user. Make a helper method that allows the user
to do it on the clouds where it is safe to do automatically.

Change-Id: I93b0c7d0b0eefdfe0fb1cd4a66cdbba9baabeb09