Remove data type enforcement on fields (cluster)
For clusters and nodes, the data type enforcement is breaking
senlinclient because we (most of the time) don't have a reasonable value
for 'created_at', 'init_at' and 'updated_at'. So those fields are left
as unassigned, thus 'null' value. However, there are problems found in
parsing a NoneType value to the enforced data type -- namely ISO8601. So
we have to remove that constraint.

Change-Id: I747ee3c2da5811c27ec67b5cf5fdd771e24d05cc