Allow resources to check their equality
The original Resource class was built atop a MutableMapping so it
included somewhat of an ability to check that two Resource instances
were the same, though it was only able to confirm equality if they
contained 100% exactly the same contents, i.e., they were probably
literally the same object. However, the newly refactored Resource class
in resource2 uses dictionaries in its composition, but doesn't inherit
from them, so it currently carries no way to check the equality of two
instances.

This introduces the ability to compare two new Resource instances to see
if they're the same by looking at the attributes stored in their
_ComponentManager instances.

This was noticed while implementing changes to the compute service while
refactoring it for resource2. The change will make a bunch of tests pass
that were depending on the ability to compare two resource instances for
equality.

Change-Id: Id6c4a5a4e26ccb8ea35900db505b43bbb19dbea9