Add temporary test_proxy_base2
In order to get the resource2 refactors moving along, the
test_proxy_base code currently points to "resource" in a lot of the
strings that get passed into various mock points. This temporarily adds
a second test_proxy_base alongside it so that we can have this refactor
move along in a per-service way. Otherwise we'd have to do one mass
change, which is far too big.

FWIW, this was originally included in the compute refactor change but
I'm splitting it out so that multiple changes can be done in parallel.

Change-Id: I7282a2413bbfd8c0e97e9b13d8141ebb30834591