Lay the groundwork for per-resource cache
There are currently two caches, the global cache, and the per-resource
batch operation caches. The global cache is managed by dogpile. The
per-resource caches are managed by hand. However, dogpile has a mode of
operation that is precisely what we do in our batched operation
per-resource caches by hand. So set ourselves up to have and use
per-resource caches.

Step one is to make the decorator accept an argument that it'll use to
look up the cache.

Change-Id: I4fbde2f4528aeaee5c23f0e1d1e4590090acb6ba