Create and return per-resource caches
As the next step in consolidating per-resource caching, we should
actually create more than one dogpile.cache region. We should only do
that if the resource is a thing we actually are going to cache (which
means it has a list_ method) Finally, we should return it if requested,
otherwise we should return the good old self._cache global cache.

Change-Id: Id0a5f65437a8bb4b911eddf54f14369d6b79d3fd