Add support for deprecating cloud profiles
We've had HP shut down their public cloud, and DreamHost has recent spun
up a new cloud with a different operational profile and auth_url that is
essentially a replacement cloud. That means people using the old
information need to do things, so we need to be able to communicate that
to them.

Add support for adding a deprecated status to a vendor profile, as well
as a verbose message explaining what the user may do to remediate.

Change-Id: I19b67d7cd71fba2d9da0e3a6adb2d229ead65396