Add API microversion support
This patch proposes a default header to be injected when sending
requests to a service. The header is added as the support to
the micro-version specification from OpenStack API WorkingGroup [1].

We can use this 'api_version' property to track the micro-versions of
service APIs we support in SDK. Users can optionally specify such an API
version request as preferences in the 'Profile' object.

[1]
http://git.openstack.org/cgit/openstack/api-wg/tree/guidelines/microversion_specification.rst

Change-Id: I421232a73de5dcc423a40acb79202a1f12bed430