Replace _transpose_component with _filter_component
The _transpose_component method needs to be adjusted as it currently
causes attribute lookups on Resource instances to differ
depending on how they were constructed. This is an oversight from the
refactoring of this class, as data is stored internally using
server-side names instead of client-side names, but the
_transpose_component method was implemented when data was being stored
in client-side naming, i.e., how the Body and Header attributes are
named on Resource classes.

This oversight happened for two reasons:

1. The Resource.list call doesn't use _transpose_component but other
calls, e.g., Resource.get, do. At least for me, I find it easier during
sanity checks against a live OpenStack cloud to use the list methods,
look at instances in the list, and check certain attributes match what I
set them to. We don't have a strong functional test suite, but that's
where this would have been caught, as we'd be able to call list and then
get each of the items individually and ensure they have the same
values. Such a test would have failed as a number of attributes would
expect values and return None when retrieved via `get` calls.

2. The manual testing I had been doing in lieu of #1 was not
sufficiently thorough, as it was generally looking at only a few
attributes, and those attributes didn't require the _transpose_component
code path. When an attribute's client-side name matches the server-side
name, everything works fine regardless of how the underlying data was
obtained. So checking Server.status will always work, but
Server.created_at depends on how you got your response.

Given the following two calls:

a = list(conn.compute.servers())[0]
b = conn.compute.get_server(a.id)

`a.created_at` would work as expected, because Resource.list doesn't
call _transpose_component. The `created_at` attribute looks for a key
"created" that came from the server-side response data. Because
Resource.list wasn't transposing the data, it was stored appropriately
as "created" and the body manager knew to look in the dictionary for
"created" when a user access `a.created_at`.

However, `b.created_at` would not work as expected and would return
None. This is because Resource.get was sending the response body through
_transpose_component, which caused the "created" key to be stored
internally as "created_at". The body manager would then look for
"created" in the dictionary when looking at the mapping of client to
server names, and it wouldn't find anything beacuse we went against the
mapping and stored as a client name. It would thusly return the default
setting for that Body instance, which is None.

An easy fix is to just remove _transpose_component and all of its
uses in the Resource class as well as one subclass. Several tests
were mocking it out so they were also updated to remove the reference.

However, as just removing allows extraneous keys into our internal
dictionaries, switching _transpose_component to more of a filter model
is a better solution. This way we only return a dictionary including
keys that we're expecting to actually work with, rather than letting
everything possible in, which a simple removal would do. One example of
a place where this approach to the fix matters is in the Keypair
resource, where we are ignoring the "id" key that is both
inconsistently available and contains inconsistent values. That
resource looks to the "name" key as both the name attribute and an
alternate id, which this change supports properly. Just removing
_transpose_component reintroduces the inconsistent `id` internally,
which then makes it possible to access it via the `id` attribute on
Keypair, which breaks the id/name separation and consistency.

Change-Id: Ida7c68c78910ee7adc9d9b73ca9cf280bc303cf5
Closes-Bug: 1600365